Separate advanced output functionality from output-level option [#47]
The advanced output is now set with the advanced-output-sites option.
It is a comma-separated list of tracker types like match-sites.
To enable it for all sites, set the option to "all".

Personal info like passkey/authkey/cookies is now censored to prevent
accidental exposure when asking for help. Please report any cases
where this is not the case.