Change method for matching release groups
Instead of capturing from the whole release name which is unreliable and inconsistent,
when the release-groups filter option is used, those values are searched for in certain
locations in the release name. Specifically in square brackets at the beginning, in square
brackets at the end, and after a hyphen and optional spaces at the end.

The consequences of this are that the release group will not be part of advanced output at this time unless
release group is set as a capture in the tracker file, which will now be rare. The release group macro will not
be populated unless a filter is matched while having release-groups set or release group is set as a capture
in the tracker file.

The benefits are that this should work out of the box for most trackers without needing to have a capture in tracker
files while retaining the ability to capture release group from the tracker file if necessary or desired.