A working method to set the Irrsi/Gui IP in the autodl.cfg
This is to support custom networking environments that deviate from the vanilla OS practices and are not currently supported.

To do this a new configuration option has been created that allows setting the IP in the autodl.cfg:

gui-server-ip = 10.0.0.1

When this setting is empty the default value is 127.0.0.1 as per the expected defaults.

This prevents compromising the default installation practice whilst providing extended and easy to document functionality.